+++ Src/WindowsServer/WindowsServer.Shared/Implementation/WindowsServerEventSource.cs
@@ -267,7 +267,7 @@
         [Event(
             28,
             Message = "Azure IMS returned field '{0}' with an invalid/unexpected value. Not adding this value to heartbeat properties.",
-            Level = EventLevel.Informational)]
+            Level = EventLevel.Warning)]
         public void AzureInstanceMetadataValueForFieldInvalid(string fieldWithInvalidValue, string applicationName = "Incorrect")
         {
             this.WriteEvent(
@@ -323,6 +323,94 @@
                 32,
                 exception ?? string.Empty,
                 innerException ?? string.Empty,
+                this.ApplicationName);
+        }
+
+        [Event(33,
+            Message = "App Services Heartbeat Provider: Failed to obtain Azure App Services environment variable '{0}'. Exception raised: {1}",
+            Level = EventLevel.Warning)]
+        public void AppServiceHeartbeatPropertyAquisitionFailed(string envVarName, string exceptionStr, string appDomainName = "Incorrect")
+        {
+            this.WriteEvent(
+                33,
+                envVarName ?? "unknown",
+                exceptionStr ?? "unknown-exception",
+                this.ApplicationName);
+        }
+
+        [Event(34,
+            Message = "App Services Heartbeat Provider: Could not obtain the Heartbeat Manager instance during initialization. Exception raised: {0}",
+            Level = EventLevel.Warning)]
+        public void AppServiceHeartbeatManagerAccessFailure(string exceptionStr, string appDomainName = "Incorrect")
+        {
+            this.WriteEvent(
+                34,
+                exceptionStr ?? "unknown-exception",
+                this.ApplicationName);
+        }
+        
+        [Event(35,
+            Message = "App Services Heartbeat Provider: Accessing the Hearbeat Manager failed as it is not in the list of available modules.",
+            Level = EventLevel.Warning)]
+        public void AppServiceHeartbeatManagerNotAvailable(string appDomainName = "Incorrect")
+        {
+            this.WriteEvent(
+                35,
+                this.ApplicationName);
+        }
+
+        [Event(36,
+            Message = "App Services Heartbeat Provider: Failed to set Azure App Services heartbeat values. Exception encountered: {0}",
+            Level = EventLevel.Warning)]
+        public void AppServiceHeartbeatPropertySettingFails(string exceptionStr, string appDomainName = "Incorrect")
+        {
+            this.WriteEvent(
+                36,
+                exceptionStr ?? "unknown-exception",
+                this.ApplicationName);
+        }
+
+        [Event(37,
+            Message = "App Services Heartbeat Provider: Request to set heartbeat properties when the heartbeat property manager is null.",
+            Level = EventLevel.Warning)]
+        public void AppServiceHeartbeatSetCalledWithNullManager(string appDomainName = "Incorrect")
+        {
+            this.WriteEvent(
+                37,
+                this.ApplicationName);
+        }
+
+        [Event(38,
+            Message = "Error occured when disposing update interval timer within EnvironmentVariableMonitor. Exception: {0}",
+            Level = EventLevel.Warning)]
+        public void EnvironmentVarMonitorFailedDispose(string exceptionMsg, string appDomainName = "Incorrect")
+        {
+            this.WriteEvent(
+                38,
+                exceptionMsg ?? "unknown-exception",
+                this.ApplicationName);
+        }
+
+        [Event(39,
+            Message = "Security exception was thrown trying to read environment variable '{0}'. Disabling environment variable monitor to avoid future security exceptions. Exception: {1}",
+            Level = EventLevel.Warning)]
+        public void SecurityExceptionThrownAccessingEnvironmentVariable(string environmentVariableName, string exceptionMsg, string appDomainName = "Incorrect")
+        {
+            this.WriteEvent(
+                39,
+                environmentVariableName,
+                exceptionMsg ?? "unknown-exception",
+                this.ApplicationName);
+        }
+
+        [Event(40,
+            Message = "Error occurred when trying to update environment variables. Exception: {0}",
+            Level = EventLevel.Warning)]
+        public void GeneralFailureOccursDuringCheckForEnvironmentVariables(string exceptionMsg, string appDomainName = "Incorrect")
+        {
+            this.WriteEvent(
+                40,
+                exceptionMsg ?? "unknown-exception",
                 this.ApplicationName);
         }